1.更换transformers新api参数类型dtype 2.milvus字段metadata空处理

+++ engines/transformers_engine.py
@@ -54,7 +54,7 @@ class TransformersLLMEngine(LLMEngine):
         model = AutoModelForCausalLM.from_pretrained(
             self.model_path,
             device_map='auto' if use_cuda else None,
-            torch_dtype=torch_dtype,
+            dtype=torch_dtype,
             trust_remote_code=True
         )
